pyproject.toml
[project]
name = "polars-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastexcel>=0.14.0",
    "google-auth>=2.40.3",
    "gspread>=6.2.1",
    "jupyter>=1.1.1",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "polars>=1.30.0",
    "yaspin>=3.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]
